refactor(orm): 调整分页查询逻辑与配置参数类型
- 修改 selectPage 方法逻辑，支持不分页时使用系统最大记录数配置
- 将 DBConfig 中 getMaxPageSize 方法返回值类型从 long 改为 int
- 优化分页参数校验逻辑，增强代码可读性
- 降级 spring-jdbc 版本以兼容当前项目环境
- 更新 mssql-jdbc 依赖注释信息

+++ framework.orm/src/main/java/work/myfavs/framework/orm/DBConfig.java
@@ -194,7 +194,7 @@ public class DBConfig {
    *
    * @return 分页时每页最大记录数
    */
-  public long getMaxPageSize() {
+  public int getMaxPageSize() {
 
     return maxPageSize;
   }
